requirements.txt
PyMuPDF
openai
python_dotenv